Fix problem with messing up terminal after editing file
Most of the Ncurses initialization code is placed in a method
Menu#init_ncurses which basically does the opposite of Menu#clean_up.
Now, Ncurses is cleaned up before editing a file and re-initialized
after. This solves the problem with the terminal printing the menu
after quitting after editing a file. However, now there is a slight
flicker/delay when a file is opened.